introduce externals folder
