+++ rbc/omnisci_backend/__init__.py
@@ -1,5 +1,4 @@
 from .numpy_funcs import *  # noqa: F401, F403
-from .omnisci_mathimpl import *  # noqa: F401, F403
 from .npy_mathimpl import *  # noqa: F401, F403
 from .omnisci_array import *  # noqa: F401, F403
 from .omnisci_column import *  # noqa: F401, F403
